requirements.txt
boto3
flask
pyyaml
requests